refactor to isolate API
