+++ pkg/openfeature/openfeature.go
@@ -1,84 +1,19 @@
 package openfeature
 
 import (
-	"sync"
-
 	"github.com/go-logr/logr"
 )
-
-type evaluationAPI struct {
-	prvder  FeatureProvider
-	hks     []Hook
-	evalCtx EvaluationContext
-	logger  logr.Logger
-	mutex
-}
 
 // api is the global evaluationAPI.  This is a singleton and there can only be one instance.
 var api evaluationAPI
 
-// init initializes the openfeature evaluation API
+// init initializes the OpenFeature evaluation API
 func init() {
 	initSingleton()
 }
 
 func initSingleton() {
-	api = evaluationAPI{
-		prvder:  NoopProvider{},
-		hks:     []Hook{},
-		evalCtx: EvaluationContext{},
-		logger:  logr.New(logger{}),
-		mutex:   &sync.RWMutex{},
-	}
-}
-
-// EvaluationOptions should contain a list of hooks to be executed for a flag evaluation
-type EvaluationOptions struct {
-	hooks     []Hook
-	hookHints HookHints
-}
-
-// HookHints returns evaluation options' hook hints
-func (e EvaluationOptions) HookHints() HookHints {
-	return e.hookHints
-}
-
-// Hooks returns evaluation options' hooks
-func (e EvaluationOptions) Hooks() []Hook {
-	return e.hooks
-}
-
-func (api *evaluationAPI) provider() FeatureProvider {
-	api.RLock()
-	defer api.RUnlock()
-	return api.prvder
-}
-
-func (api *evaluationAPI) setProvider(provider FeatureProvider) {
-	api.Lock()
-	defer api.Unlock()
-	api.prvder = provider
-	api.logger.V(info).Info("set global provider", "name", provider.Metadata().Name)
-}
-
-func (api *evaluationAPI) setEvaluationContext(evalCtx EvaluationContext) {
-	api.Lock()
-	defer api.Unlock()
-	api.evalCtx = evalCtx
-	api.logger.V(info).Info("set global evaluation context", "evaluationContext", evalCtx)
-}
-
-func (api *evaluationAPI) setLogger(l logr.Logger) {
-	api.Lock()
-	defer api.Unlock()
-	api.logger = l
-	api.logger.V(info).Info("set global logger")
-}
-
-func (api *evaluationAPI) hooks() []Hook {
-	api.RLock()
-	defer api.RUnlock()
-	return api.hks
+	api = NewEvaluationAPI()
 }
 
 // SetProvider sets the global provider.
